`getSquadCommunity` moved to correct module.
