+++ src/engine/core/utils/community.ts
@@ -1,5 +1,7 @@
 import { registry } from "@/engine/core/database";
+import { Squad } from "@/engine/core/objects/squad";
 import { isStalker } from "@/engine/core/utils/class_ids";
+import { squadCommunityByBehaviour } from "@/engine/lib/constants/behaviours";
 import { communities, TCommunity } from "@/engine/lib/constants/communities";
 import {
   GameObject,
@@ -9,6 +11,14 @@ import {
   ServerHumanObject,
   TNumberId,
 } from "@/engine/lib/types";
+
+/**
+ * @param squad - target squad to get community for
+ * @returns squad community section
+ */
+export function getSquadCommunity(squad: Squad): TCommunity {
+  return squadCommunityByBehaviour.get(squad.faction);
+}
 
 /**
  * Returns community of provided object.
