+++ src/engine/core/utils/relation/relation_get.ts
@@ -3,21 +3,13 @@ import { relation_registry } from "xray16";
 import { getServerObjectByStoryId, registry } from "@/engine/core/database";
 import type { Squad } from "@/engine/core/objects/squad/Squad";
 import { assert } from "@/engine/core/utils/assertion";
+import { getSquadCommunity } from "@/engine/core/utils/community";
 import { LuaLogger } from "@/engine/core/utils/logging";
 import { EGoodwill, ERelation } from "@/engine/core/utils/relation/relation_types";
-import { squadCommunityByBehaviour } from "@/engine/lib/constants/behaviours";
 import { communities, TCommunity } from "@/engine/lib/constants/communities";
 import { GameObject, Optional, TCount, TNumberId, TRelationType, TStringId } from "@/engine/lib/types";
 
 const logger: LuaLogger = new LuaLogger($filename);
-
-/**
- * @param squad - target squad to get community for
- * @returns squad community section
- */
-export function getSquadCommunity(squad: Squad): TCommunity {
-  return squadCommunityByBehaviour.get(squad.faction);
-}
 
 /**
  * Get relation type between objects with safe null check.
