+++ src/engine/core/utils/squad/squad_get.test.ts
@@ -49,6 +49,4 @@ describe("squad utils", () => {
     expect(getObjectSquadByObjectId(gameObject.id())).toBeNull();
     expect(getObjectSquadByObjectId(serverObject.id)).toBeNull();
   });
-
-  it.todo("getSquadCommunity should correctly get community for squads");
 });
